% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calc_BD}
\alias{calc_BD}
\title{Calculate BD}
\usage{
calc_BD(data, g_control_formula)
}
\arguments{
\item{data}{a data.frame with all variables in the g_control_formula,
outcome, treated, id, and time}

\item{g_control_formula}{control formula with group level variable names}
}
\value{
BD
}
\description{
Calculate BD
}
