% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calc_controlled_beta_weights}
\alias{calc_controlled_beta_weights}
\title{Calculates Controlled Betas and Weights}
\usage{
calc_controlled_beta_weights(data, g_control_formula)
}
\arguments{
\item{data}{a data.frame}

\item{g_control_formula}{formula with group/timelevel control variables}
}
\value{
a list with two elements
\item{s_kl}{the weight given to the dyad kl's estimate}
\item{beta_hat_d_bkl}{the covariate adjusted 2x2 estimate for the dyad kl}
}
\description{
Wrapper function for finding the dyad beta/weight in the controled 
 decomposition
}
