% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calc_pgjtile}
\alias{calc_pgjtile}
\title{Calculate pgjtile and Rsq}
\usage{
calc_pgjtile(data, g_control_formula)
}
\arguments{
\item{data, }{a data.frame with the variables in the control formula and 
Dtilde}

\item{g_control_formula}{control formula with group level variables}
}
\value{
a list with two elements: 
\item{data}{updated data.frame with new variable pgjtilde}
\item{Rsq}{Rsq from the regression to generate pgjtilde}
}
\description{
Calculate pgjtile and Rsq
}
