% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calc_VD}
\alias{calc_VD}
\title{Calculate Treatment Variance}
\usage{
calc_VD(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a list of two items:
\item{data}{a data.frame with the new demeaned treatment variable}
\item{VD}{the varince of the demeaned treatment variable}
}
\description{
Calculate Treatment Variance
}
