% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calc_Vdp}
\alias{calc_Vdp}
\title{Calculate Vdp}
\usage{
calc_Vdp(data)
}
\arguments{
\item{data}{a data.frame with the variables g_p, id, time, and ptilde}
}
\value{
list with two elements
\item{data}{data.frame with new variables ptilde and dp}
\item{Vdp}{Variance of dp}
}
\description{
Calculate Vdp
}
