% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_22_kl}
\alias{calculate_beta_hat_22_kl}
\title{Calculate beta_hat_22_kl}
\usage{
calculate_beta_hat_22_kl(data)
}
\arguments{
\item{data}{}
}
\value{
beta_hat_22_kl
}
\description{
Calculate beta_hat_22_kl
}
