% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_b}
\alias{calculate_beta_hat_b}
\title{CalculateBetween Estimate}
\usage{
calculate_beta_hat_b(data)
}
\arguments{
\item{data}{a data.frame with the columns outcome and d_kt_til}
}
\value{
beta_hat_w, the between estimate
}
\description{
CalculateBetween Estimate
}
