% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_d_bkl}
\alias{calculate_beta_hat_d_bkl}
\title{Calculate beta_hat_d_bkl}
\usage{
calculate_beta_hat_d_bkl(data, treated_group, untreated_group)
}
\arguments{
\item{data}{}
}
\value{
beta_hat_d_bkl
}
\description{
Calculate beta_hat_d_bkl
}
