% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_p_bkl}
\alias{calculate_beta_hat_p_bkl}
\title{Calculate beta_hat_p_bkl}
\usage{
calculate_beta_hat_p_bkl(data)
}
\arguments{
\item{data}{}
}
\value{
beta_hat_p_bkl
}
\description{
Calculate beta_hat_p_bkl
}
