% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_w}
\alias{calculate_beta_hat_w}
\title{Calculate Within Estimate}
\usage{
calculate_beta_hat_w(data)
}
\arguments{
\item{data}{a data.frame with the columns outcome and d_ikt_til}
}
\value{
beta_hat_w, the within estimate
}
\description{
Calculate Within Estimate
}
