% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_Omega}
\alias{calculate_Omega}
\title{Calculate Omega}
\usage{
calculate_Omega(data)
}
\arguments{
\item{data}{a data frame with the columns d_ikt_til and d_it_til

return Omega}
}
\description{
Calculate the weight for the within estimate in the controlled decompositon
}
