% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_one_minus_Omega}
\alias{calculate_one_minus_Omega}
\title{Calculate 1 - Omega}
\usage{
calculate_one_minus_Omega(data)
}
\arguments{
\item{data}{a data frame with the columns d_kt_til and d_it_til}
}
\description{
Calculate the weight for the between estimates
}
