% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_one_minus_Sigma}
\alias{calculate_one_minus_Sigma}
\title{Calculate 1 - Sigma}
\usage{
calculate_one_minus_Sigma(data)
}
\arguments{
\item{data}{}
}
\value{
one_minus_Sigma
}
\description{
Calculate 1 - Sigma
}
