% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_s_kl}
\alias{calculate_s_kl}
\title{Calculate Weight (Controlled)}
\usage{
calculate_s_kl(N, Rsq, VD, Vdp)
}
\arguments{
\item{N}{observations in dyad}

\item{Rsq}{R squared from the regression in calc_pgjtilde}

\item{VD}{variance of the demeaned treatment indicator (in the dyad)}

\item{Vdp}{variance of dp (pgjtilde - ptilde)}
}
\value{
the weight given to the dyad's estimate
}
\description{
Calculate Weight (Controlled)
}
