% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate Weights for 2x2 Grid}
\usage{
calculate_weights(data, treated_group, untreated_group)
}
\arguments{
\item{data}{a data.frame}

\item{treated_group}{the identifier of the treated group}

\item{untreated_group}{the identifier of the untreated group}
}
\value{
Scalar weight for 2x2 grid
}
\description{
Calculated weights using:
 n_u - observations in untreated group
 n_t - observations in treated group
 p_t - proportion of the time treated group was treated
 n_e - observations in early treated group
 n_l - observations in late treated group
 p_e - proportion of the time early treated group was treated
 p_l - proportion of the time late treated group was treated
}
