% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_weights_controled}
\alias{calculate_weights_controled}
\title{Calculate Weights in Controlled Decomposition}
\usage{
calculate_weights_controled(data, treated_group, untreated_group, V_db)
}
\description{
Calculate Weights in Controlled Decomposition
}
