% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{castle}
\alias{castle}
\title{Data from Cheng and Hoekstra (2013, JHR)}
\format{A data.frame with 520 observations and 159 variables}
\usage{
castle
}
\description{
Data from Cheng and Hoekstra (2013, JHR)
}
\keyword{datasets}
