% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{collapse_x_p}
\alias{collapse_x_p}
\title{Collapse Control Variable and Predicted Treatment}
\usage{
collapse_x_p(data, control_formula)
}
\arguments{
\item{data}{a data.frame}

\item{control_formula, }{a formula}
}
\value{
a list of two elements:
\item{data}{updated data frame with new treatment group/time level variables} 
\item{g_control_formula}{the control formula but with group level variable 
 names}
}
\description{
Collapse controll variables and predicted treatment to treatment group/time 
 level.
}
