% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{create_treatment_groups}
\alias{create_treatment_groups}
\title{Create Grid of Treatment Groups}
\usage{
create_treatment_groups(data, control_vars, return_merged_df = FALSE)
}
\arguments{
\item{data}{a data.frame used to create groups - MUST obey naming convention 
used in `bacon()`. i.e. columns are ["id", "time", "outcome", "treated"]}

\item{control_vars}{list of control variables}

\item{return_merged_df}{Defaults to `FALSE` whether to return merged data
as well as grid of treatment groups.}
}
\value{
If return_merge_df is `TRUE` then a list with two elements:
\item{two_by_twos}{data.frame with all treatment group dyads and NA 
 weight/estimate columns}
\item{data}{updated data.frame with the new variable treat_time}

If return_merge_df is `FALSE` then only the two_by_twos data.frame is 
 returned
}
\description{
Create Grid of Treatment Groups
}
