% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divorce}
\alias{divorce}
\title{Data from Stevenson and Wolfers (2006, QJE)}
\format{A data.frame with 3366 observations and 147 variables}
\usage{
divorce
}
\description{
Data from Stevenson and Wolfers (2006, QJE)
}
\keyword{datasets}
