% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_reform}
\alias{math_reform}
\title{Aggregated data from Goodman (In Press)}
\format{A data.frame with 520 observations and 5 variables
\describe{
  \item{state}{the state}
  \item{class}{the high school class}
  ...
}}
\usage{
math_reform
}
\description{
A data set comntaining state/year level data on an educational reform and
future income.
}
\keyword{datasets}
