% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{run_fwl}
\alias{run_fwl}
\title{Run Frisch-Waugh-Lowell Regression}
\usage{
run_fwl(data, control_formula)
}
\arguments{
\item{data}{a data.frame}

\item{control_formula}{a fomula}
}
\value{
data, a data.frame with predictions/residuals (and their
 transfirmations) from FWL regession
}
\description{
Predict treatment using time varing covariates and calculates
}
