% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{scale_weights}
\alias{scale_weights}
\title{Scale Weights}
\usage{
scale_weights(two_by_twos)
}
\arguments{
\item{two_by_twos}{data.frame containing 2x2 weights}
}
\value{
two_by_twos data frame with resaled weight column
}
\description{
scale_weights scales the two_by_two weights so that they sum to 1
}
