% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{subset_data}
\alias{subset_data}
\title{Subset Data}
\usage{
subset_data(data, treated_group, untreated_group)
}
\arguments{
\item{untreated_group}{}
}
\value{
subsetted data.frame
}
\description{
Subset Data
}
