% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{unpack_variable_names}
\alias{unpack_variable_names}
\title{Unpack Variable Names from Formula}
\usage{
unpack_variable_names(formula)
}
\arguments{
\item{formula}{formula call from bacon()}
}
\value{
A list with 3 elements: 
\item{outcome_var}{name of the outcome variable from the LHS of formula}
\item{treated_var}{name of the treatment indicator variable, the first 
 variable from the RHS of formula}
\item{control_vars}{All variables on RHS of formula except the treated_var. 
 Only need to know if length(control_vars) > 0}
}
\description{
Unpack Variable Names from Formula
}
