% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{bacon}
\alias{bacon}
\title{Goodman-Bacon Decomposition}
\usage{
bacon(formula, data, id_var, time_var)
}
\arguments{
\item{formula}{a symbolic representation of the
model to be fitted. Must be of the form y ~ D + ..., where D is the binary 
treatment indicator.}

\item{data}{a data.frame containing the variables in the model.}

\item{id_var}{character, the name of id variable for units.}

\item{time_var}{character, the name of time variable.}
}
\value{
data.frame of all 2x2 estimates and weights
}
\description{
bacon() is a function that perfroms the Goodman-Bacon decomposition for
 differences-in-differences with variation in treatment timing.
}
\examples{
# Castle Doctrine -----------------------------------------------------------
df_bacon <- bacon(l_homicide ~ post,
                  data = bacon::castle,
                  id_var = "state",
                  time_var = "year")

\donttest{

library(ggplot2)
ggplot(df_bacon) +
  aes(x = weight, y = estimate, shape = factor(type)) +
  labs(x = "Weight", y = "Estimate", shape = "Type") +
  geom_point()

  }

}
\author{
Evan Flack
}
