% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{bacon}
\alias{bacon}
\title{Decompose two-way fixed effects models}
\usage{
bacon(df, id_var = "id", time_var = "time",
  treat_time_var = "treat_time", treated_var = "treated",
  outcome_var = "outcome")
}
\arguments{
\item{df}{a data.frame}

\item{id_var}{character, name of id variable for units}

\item{time_var}{character, name of time variable}

\item{treat_time_var}{character, name of variable indicating when treatment
switched on for unit}

\item{treated_var}{character, name of treatment variable (binary)}

\item{outcome_var}{character, name of outcome variable}
}
\value{
data.frame of all 2x2 estimates and weights
}
\description{
bacon() is a function that decomposes two-way fixed effects models following
Goodman-Bacon into all 2x2 difference in difference estimates
}
\examples{
math_reform <- bacon::math_reform
math_reform[is.na(math_reform$reformyr_math), "reformyr_math"] <- 99999

bacon(math_reform, id_var = "state", time_var = "class",
      treat_time_var = "reformyr_math",  treated_var = "reform_math",
     outcome_var = "incearn_ln")

}
