% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{bacon}
\alias{bacon}
\title{Goodman-Bacon decomposition}
\usage{
bacon(df, id_var = "id", time_var = "time", treated_var = "treated",
  outcome_var = "outcome")
}
\arguments{
\item{df}{a data.frame}

\item{id_var}{character, name of id variable for units}

\item{time_var}{character, name of time variable}

\item{treated_var}{character, name of treatment variable (binary)}

\item{outcome_var}{character, name of outcome variable}
}
\value{
data.frame of all 2x2 estimates and weights
}
\description{
bacon() is a function that perfroms the Goodman-Bacon decomposition for
 differences-in-differences with variation in treatment timing.
}
\examples{
df_bacon <- bacon(df = bacon::math_reform,
                  id_var = "state",
                  time_var = "class",
                  treated_var = "reform_math",
                  outcome_var = "incearn_ln")
ggplot(df_bacon) +
  aes(x = weight, y = estimate, shape = factor(type)) +
  labs(x = "Weight", y = "Estimate", shape = "Type") +
  geom_point()
}
