% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calc_VD}
\alias{calc_VD}
\title{Calculate Treatment Variance}
\usage{
calc_VD(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a list of data (a data.frame) and VD, the varince of the demeaned
 treatment variable
}
\description{
Calculate Treatment Variance
}
