% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_b}
\alias{calculate_beta_hat_b}
\title{Calculate between beta hat}
\usage{
calculate_beta_hat_b(data)
}
\arguments{
\item{data}{}
}
\value{
beta_hat_b
}
\description{
Calculate between beta hat
}
