% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_beta_hat_d_bkl}
\alias{calculate_beta_hat_d_bkl}
\title{Calculate Covariate Adjusted Between Beta Hat}
\usage{
calculate_beta_hat_d_bkl(Rsq, VD, BD, Vdp, Bb)
}
\arguments{
\item{Rsq}{R squared from the regression in calc_pgjtilde}

\item{VD}{variance of the demeaned treatment indicator (in the dyad)}

\item{BD}{coeffieicient on treatment from a regression of the outcome on 
treatment and group level control variables with unite and time FE}

\item{Vdp}{variance of dp (pgjtilde - ptilde)}

\item{Bb}{coefficient on dp from a regression of the outcome on dp with unit 
and time fixed effects}
}
\value{
dyad's covariate adjusted between estimate
}
\description{
Calculate Covariate Adjusted Between Beta Hat
}
