% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_ps}
\alias{calculate_ps}
\title{Calculate "p's"}
\usage{
calculate_ps(data, control_formula)
}
\arguments{
\item{data}{}

\item{control_formula}{#' @return data with new p variables}

\item{ca}{data.frame}
}
\description{
Calculate "p's"
}
