% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate Weights for 2x2 Grid (uncontrolled)}
\usage{
calculate_weights(data, treated_group, untreated_group)
}
\arguments{
\item{data}{a data.frame}

\item{treated_group}{the identifier of the treated group}

\item{untreated_group}{the identifier of the untreated group}
}
\value{
Scalar weight for 2x2 grid
}
\description{
Calculated weights for the uncontrolled decomposition using:
 n_u - observations in untreated group,
 n_k - observations in earlier treated group,
 n_l - observations in later treated group,
 D_k - proportion of time the earlier treated group was treated,
 D_l - proportion of time the later treated group was treated.
}
