% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{castle}
\alias{castle}
\title{Data from Cheng and Hoekstra (2013, JHR)}
\format{A data.frame with 520 observations and 159 variables
\describe{
  \item{st}{The state (unit of analysis).}
  \item{year}{Calendar year (time).}
  \item{l_homicide}{Log of state/year homicide rate}
  \item{post}{Indicator whether castle reform has been implemented}
}}
\usage{
castle
}
\description{
Data from Cheng and Hoekstra (2013, JHR)
}
\keyword{datasets}
