% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{collapse_x_p}
\alias{collapse_x_p}
\title{Collapse Xs and Predicted Treatment}
\usage{
collapse_x_p(data, control_formula)
}
\arguments{
\item{data}{a data.frame}

\item{control_formula, }{a formula}
}
\value{
a list of data (a data.frame) and g_control_formula, the formula with
 the group level variable names
}
\description{
Collapse Xs and predicted treatment to treatmet group/time level.
}
