% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{create_treatment_groups}
\alias{create_treatment_groups}
\title{Create Grid of Treatment Groups}
\usage{
create_treatment_groups(data, control_vars, return_merged_df = FALSE)
}
\arguments{
\item{data}{dataset used to create groups - MUST obey naming convention used
in `bacon()`. i.e. columns are ["id", "time", "outcome", "treated"]}

\item{return_merged_df}{Defaults to `FALSE` whether to return merged data
as well as grid of treatment groups.}
}
\value{
data.frame describing treatment groups and empty weight and estimate
column set to 0.
}
\description{
Create Grid of Treatment Groups
}
