% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename Variables}
\usage{
rename_vars(data, id_var, time_var, outcome_var, treated_var)
}
\arguments{
\item{data}{a data.frame}

\item{id_var}{character, name of id variable}

\item{time_var}{character, name of time variable}

\item{outcome_var}{character, name of outcome variable}

\item{treated_var}{character, name of binary treatment variable}
}
\value{
data.frame with renmaed columns
}
\description{
Rename Variables
}
