% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename Variables}
\usage{
rename_vars(data, id_var, time_var, outcome_var, treated_var)
}
\arguments{
\item{data}{a data.frame}

\item{treated_var}{}
}
\value{
data.frame with renmaed columns
}
\description{
Rename Variables
}
