% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{run_fwl}
\alias{run_fwl}
\title{Run Frisch-Waugh-Lowell Regression}
\usage{
run_fwl(data, control_formula)
}
\arguments{
\item{data}{a data.frame}

\item{control_formula}{a fomula}
}
\value{
updated data.frame with predictions/residuals (and their
 transformations) from the FWL regession
}
\description{
Predicts treatment using time-varing covariates and calculates predicted 
 values/residuals
}
