% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{unpack_variable_names}
\alias{unpack_variable_names}
\title{Unpack Variable Names from Formula}
\usage{
unpack_variable_names(formula)
}
\arguments{
\item{formula}{formula call from bacon.}
}
\value{
a list with 3 elements: outcome_var, treated_var, control_vars
}
\description{
Unpack Variable Names from Formula
}
